56_DAY_PRACTICE/day10_cloud_storage/fastapi-gemini/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-genai==1.64.0
pydantic==2.12.0
python-dotenv==1.1.1
